# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CareNexa API
  version: 0.1.0
  description: >
    Local REST API for the CareNexa clinic CRM. Insights are rule-based.
    Money fields are decimal strings in Philippine pesos.
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      security: []
      summary: Sign in
      responses:
        '200': { description: Token and user }
        '401': { description: Invalid credentials }
...
  /api/customers:
    get:
      summary: List customers
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        '200': { description: Paged customers }
    post:
      summary: Create a customer
      responses:
        '201': { description: Created }
  /api/customers/{id}:
    get:
      summary: Customer profile
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Profile }
    put:
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
...
  /api/campaigns/{id}/activate:
    post:
      summary: Simulate delivery to the target segment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job result }
...